chore(conventions): Refactor code to follow naming conventions
- Renamed class files to follow Pascal case naming convention
- Refactored error message of the error handling middleware

165237834
